+++ src/demo.js
@@ -25,8 +25,6 @@ modify_hljs(function(hljs){
 ReactDOM.render(
   <Valine  appId={"I5DAxOhp2kPXkbj9VXPyKoEB-gzGzoHsz"}
            appKey={"lGPcHd7GL9nYKqBbNEkgXKjX"}
-    // requireEmail={true}
-    // lang={"en"}
            canBeModify={true}
            placeholder={"Try“:” to input emoji"}
            nest={true}