+++ src/index.js
@@ -22,12 +22,10 @@ modify_hljs(function(hljs){
 ReactDOM.render(
   <Valine  appId={"I5DAxOhp2kPXkbj9VXPyKoEB-gzGzoHsz"}
            appKey={"lGPcHd7GL9nYKqBbNEkgXKjX"}
-           // requireEmail={true}
-           // lang={"en"}
            canBeModify={true}
            placeholder={"Try“:” to input emoji"}
            nest={true}
-           nestLayers={1}
+           CommentClass={"Comment_demo"}
   >
     <App />
   </Valine>, document.getElementById('root'));