+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+##### 0.5.2
+
+* 修复提交后，`textarea`不能恢复高度的bug
+* 修改自定义`leancloud`上储存**评论**的Class默认名称为`"Comment"`(兼容之前版本和邮件回复功能)
+* 增加自定义`leancloud`上储存**页面阅读量**的Class名称，默认为`"Counter"`
+
 ##### 0.5.1 (2020-3-15)
 
 * `API`请求替换`leancloud-sdk`，显著减少包大小(parsed size from `479.2kb` to `331.92kb`)